requirements.txt
Flask==2.2.5
mysql-connector-python==8.1.0
Werkzeug==2.2.3
python-dotenv==1.0.0
